requirements.txt
openai>=1.0.0
piexif>=1.1.3
watchdog>=3.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
